Pixelmon V1.1.7.12b
A lot of mod updates, config changes, hot fix to the nether portal not working and 40% of quests been added, a lot of the quests does not have rewards, a rewards table will be created and will be added to the quests with a random item reward.

Check the changelogs for changes.

Added more mods and removed some mods that caused issues.

Co-Authored-By: reconsolidate <58376584+reconsolidate@users.noreply.github.com>